# RFC: Speedup generic LowerTriangular chol