The generic `LowerTriangular` `_chol!` function is a bit slower than its `UpperTriangular` counterpart. After removing some unnecessary multiplications and branches they are about the same speed. Bellow are a couple of benchmarks to demonstrate. On master
```julia
srand(1234)
A = randn(100,100) |> t -> t't
@benchmark invoke(Base.LinAlg._chol!, Tuple{AbstractMatrix, Type{LowerTriangular}}, copy($A), LowerTriangular)
@benchmark invoke(Base.LinAlg._chol!, Tuple{AbstractMatrix, Type{UpperTriangular}}, copy($A), UpperTriangular)
```
Gives mean timings of 177.964 μs for the `LowerTriangular` version and 142.604 μs for the `UpperTriangular`. With this PR the same benchmark gives mean timings of 143.273 μs and 142.406 μs respectively.