@bshall 2017-06-05
Thanks for the feedback. I've made the suggested changes.