@KristofferC 2017-06-05
For large matrices this is slower (likely due to worse access pattern).

For example, for a 1000 x 1000 matrix I get 180 ms with original code and 257 ms with this code.