@andreasnoack 2017-06-05
@KristofferC Thanks for checking a larger matrix. I think it might be possible to improve the original version without reordering the loops.