@bshall 2017-06-05
@KristofferC Thanks for pointing that out. Another option (the latest commit) is to keep the original loop order but move the scaling to the end. This requires an extra loop but the performance is still pretty good. I've included some more detailed benchmarks below.

| size | master/lower | master/upper | option 1 | option 2 |
| --- | --- | --- | --- | --- |
| 10 | 564.525 ns | 379.070 ns | 387.591 ns | 462.354 ns |
| 100 | 178.669 μs | 146.599 μs | 152.830 μs | 157.919 μs |
| 1000 | 151.652 ms | 144.080 ms | 582.975 ms | 124.658 ms |
| 2000 | 1.333 s | 1.316 s | 2.320 s | 1.112 s |