@KristofferC 2017-06-05
Perhaps you can try again with my `@simd` suggestion?