@bshall 2017-06-05
@KristofferC I've added your `@simd` to the benchmarks table and it does indeed make a significant difference for larger matrices.