@KristofferC 2017-06-21
If the SIMD version was faster it should be added.