@bshall 2017-06-22
I've updated the lower triangular version to use SIMD. Does anyone have any idea why SIMD won't kick in on the upper triangular version though?